refactor(cli): move reporter module under reporting
