+++ crates/tsz-cli/src/reporting/reporter.rs
